requirements.txt
fasttext-wheel==0.9.2
hanlp==2.1.0a53
grpcio==1.34.1
tensorflow==2.5.0
tensorflow-serving-api==2.5.1